fix tests (race condition with PAPI_init)
